Player Hand drill added, pick up ore, and some tile texturing
Hand drill works great and splits ore into two categories: Common and
Enriched. This way the Resource Manager shows the sum of both categories
as Ore, but when that Ore is ready to be refined it'll output either
Common or Enriched product.
Tile texture still working on it, just got all the tile graphics from
Jason.