Extraction Building base class
- New logic for extraction buildings makes it easier to pass resources
from one to another (like for a Storage Unit and Farms) and they all run
using the same set of Coroutines. Pretty cool! I want to expand the area
it searches for the resource it wants to extract to that water pumps can
find water even if they are more than 1 tile away from it. Yay!