Ship Inventory and Storage continued
-Storage works great. You set the output of a building by clicking on
the set input button of a receiving unit (Farm, and Storage Units).
Sounds redundant but it works well because the receiving building isn't
setting anything so it can receive multiple inputs by setting unique
outputs on all the buildings sending to them.

Ship inventory is at its most basic right now. When Storage Units become
full they will automatically beam their entire contents to the ship. The
ship right now is just holding a bunch of int with a simple receive
method.
Need to work on separating ore into common and rich and keeping all that
information safe as we load from a world level to the ship.