Enemy Jump Attacks
- Made enemies path towards towers that hit them (or player) and jump at
them as if they are hitting them. It looks nice. But with this added
effect a bunch more issues popped up.
- Spawned Tiles were not always pointing to the right object. When
placing a building the array will need to know that for every tile that
gameobject takes up in rendered space, the Spawned Tiles array needs to
point to that same gameobject. This is fixed now, hopefully it won't
break!!!
- Now mouse controller points to the tile and gameobject under the
mouse. Since all tiles that represent an object are pointing to the same
gameobject, this helps for destroying and detecting when the mouse is
over a building.

- Pathfinding: enemies needed a way to escape from a position when
either the tile they are currently trying to walk on or the target tile
(end of path) is not a walkable tile. To escape we find the tile around
the unit that IS walkable and still has the closest distance to the
target.