Worked on Temporary Inventory
- Temp Inventory keeps track of all resources gathered during a session
on the planet. Those resources don't actually get beamed to ship until
the player Launches / Transports using the Launchpad, back to the ship.
BUT! The Launchpad is locked while the Terraformer is still working,
forcing the player to finish the Terraforming stages before they can
Launch back to the ship with their new loot/gatherings.
This makes the time on the planet already feel a lot more pressure
intensive, which is what I'm looking for. Add better balanced units and
an aggressive AI and it can really get there!!!