Climate Manager
When launching from the ship the Player chooses what Climate they wish
to land in (this will later be limited by progression/ split into
zones).
The climate affects map generation when choosing the land type
combinations to use for the island's tiles.

Also created a Rock Generator to keep the code for that separate from
ResourceGrid. This still needs work with finding the right way to
generate formations (still getting positions that are on the water or
too close to shore)