Pathfinding - guaranteed legal path target
For a while I've had the issue that any building/tower targets were
always un-walkable tiles and enemy units were unable to get a path to
them w/o some sort of offset.
Using Lerp I've found a way better solution that returns a more
"natural" offset because it relates specifically to the direction the
unit is moving/pathing towards.