Adv. Build Controls III
Got the position to work right and I can drag and build.
I want to now check to see if a building uses draggable building
controls or not so that I can still use the tile placing logic of
multiple tiles according to sprite size. Only elements that are of a
sprite size of no more than 1 tile can be draggable.